int cast

+++ src/LidarDriver.cpp
@@ -8,7 +8,7 @@
 
 // CONSTRUCTOR
 LidarDriver::LidarDriver(double resolution)
-    : angular_resolution_{resolution}, buffer_(BUFFER_DIM, std::vector<double>((RANGE / angular_resolution_) + 1))
+    : angular_resolution_{resolution}, buffer_(BUFFER_DIM, std::vector<double>(static_cast<int>(RANGE / angular_resolution_ + 1)))
 {
     if (angular_resolution_ < 0.1 || angular_resolution_ > 1.0)
         throw std::invalid_argument("angular resolution must be [0.1°, 1.0°]");
